backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Hackathon Todo Evolution - FastAPI Backend"
requires-python = ">=3.13.0"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "pydantic>=2.10.0",
    "python-jose[cryptography]>=3.3.0",
    "pyjwt>=2.9.0",  # JWT verification with JWKS support
    "cryptography>=46.0.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.30.0",
    "psycopg>=3.2.0",
    "psycopg[binary]>=3.2.0",
    "email-validator>=2.0.0",
    "openai>=1.0.0",
    "fastmcp>=0.1.0",
    "slowapi>=0.1.9",  # Rate limiting
    "structlog>=24.0.0",  # Structured logging
    "alembic>=1.13.0",  # Database migrations
    "qdrant-client>=1.12.0",  # Vector database for chat history embeddings
    "sentence-transformers>=3.0.0",  # Embeddings for semantic search
    "python-dateutil>=2.8.0",  # RRULE parsing for recurring tasks
    "apscheduler>=3.10.0",  # Background jobs for due date reminders
    "dapr==1.14.0",  # Dapr Python SDK (pinned version per NFR)
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",  # Mock support for subscriber testing (T110, T160)
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
